cloud_server/requirements.txt
fastapi
uvicorn
python-multipart
transformers
Pillow
torch
motor
python-dotenv
jinja2